+++ libraries/CHT832X/CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.3.1] - 2025-11-21
+- fix #7, update readme.md about **getNIST()**
+- add **CHT8320** and **CHT8325** as derived classes (convenience).
+- add example CHT832X_plotter.ino
+- minor edits
+
 ## [0.3.0] - 2025-10-03
 - fix #5, temperature above ~40C flipped to negative
 - made all floats explicit (0.0f iso 0.0)